+++ src/SkyDolly/src/ErrorCodes.h
@@ -22,42 +22,16 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
-#ifndef TERMINATIONDIALOG_H
-#define TERMINATIONDIALOG_H
+#ifndef ERRORCODES_H
+#define ERRORCODES_H
 
-#include <QDialog>
-#include <QTimer>
-
-class QString;
-class QTextStream;
-
-namespace Ui {
-class TerminationDialog;
-}
-
-#include "../UserInterfaceLib.h"
-
-class USERINTERFACE_API TerminationDialog : public QDialog
+namespace ErrorCodes
 {
-    Q_OBJECT
-public:
-    TerminationDialog(const QString title, const QString reason, const QString stackTrace, QWidget *parent = nullptr);
-    virtual ~TerminationDialog();
-
-private:
-    Ui::TerminationDialog *ui;
-    QString m_title;
-    QString m_reason;
-    QString m_stackTrace;
-
-    void initUi() noexcept;
-    void frenchConnection() noexcept;
-    QString createReport() const noexcept;
-    void enumeratePluginContent(const QString &pluginDirectoryPath, QTextStream &out) const;
-
-private slots:
-    void copyReportToClipboard() noexcept;
-    void createIssue() const noexcept;
+    constexpr int Ok {0};
+    constexpr int StandardException {-1};
+    constexpr int UnknownException {-2};
+    constexpr int Signal {-3};
+    constexpr int UnknownError {-99};
 };
 
-#endif // TERMINATIONDIALOG_H
+#endif // ERRORCODES_H