+++ src/SkyDolly/src/ExceptionHandler.h
@@ -30,14 +30,19 @@
 
 #include <QString>
 
+/**
+ * The exception handler handles unexpected application exceptions, due to
+ * - Uncaught exceptions
+ * - Unexpected exceptions (thrown from "noexcept" methods)
+ * - Fatal signals such as "segmentation faults"
+ */
 class ExceptionHandler
 {
 public:
-    static void handle(const QString &title, const QString &stackTrace, const std::exception &ex) noexcept;
-    static void handle(const QString &title, const QString &stackTrace, const QString &exceptionMessage) noexcept;
+    static void handleError(const QString &title, const QString &stackTrace, const std::exception &ex) noexcept;
+    static void handleError(const QString &title, const QString &stackTrace, const QString &exceptionMessage) noexcept;
     static void handleTerminate() noexcept;
-    static void signalHandler(int signal) noexcept;
-    static int getSignal() noexcept;
+    static void handleSignal(int signal) noexcept;
 private:
     static QString errorCodeToString(const std::error_code &code);
     static QString exceptionToString(const std::exception &ex);