+++ src/SkyDolly/src/SignalHandler_Windows.cpp
@@ -22,42 +22,19 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  */
-#ifndef TERMINATIONDIALOG_H
-#define TERMINATIONDIALOG_H
+#include <memory>
 
-#include <QDialog>
-#include <QTimer>
+#include "SignalHandler.h"
 
-class QString;
-class QTextStream;
+struct SignalHandlerPrivate
+{};
 
-namespace Ui {
-class TerminationDialog;
-}
+// PUBLIC
 
-#include "../UserInterfaceLib.h"
+SignalHandler::SignalHandler() = default;
+SignalHandler::~SignalHandler() = default;
 
-class USERINTERFACE_API TerminationDialog : public QDialog
+void SignalHandler::registerSignals() noexcept
 {
-    Q_OBJECT
-public:
-    TerminationDialog(const QString title, const QString reason, const QString stackTrace, QWidget *parent = nullptr);
-    virtual ~TerminationDialog();
-
-private:
-    Ui::TerminationDialog *ui;
-    QString m_title;
-    QString m_reason;
-    QString m_stackTrace;
-
-    void initUi() noexcept;
-    void frenchConnection() noexcept;
-    QString createReport() const noexcept;
-    void enumeratePluginContent(const QString &pluginDirectoryPath, QTextStream &out) const;
-
-private slots:
-    void copyReportToClipboard() noexcept;
-    void createIssue() const noexcept;
-};
-
-#endif // TERMINATIONDIALOG_H
+    // No POSIX signals or similar on Windows
+}